modified:   src/app/app.module.ts 	renamed:    src/app/pages/previsiones/modal-locales/modal-locales.component.html -> src/app/pages/previsiones/old- modal-locales/modal-locales.component.html 	renamed:    src/app/pages/previsiones/modal-locales/modal-locales.component.scss -> src/app/pages/previsiones/old- modal-locales/modal-locales.component.scss 	renamed:    src/app/pages/previsiones/modal-locales/modal-locales.component.spec.ts -> src/app/pages/previsiones/old- modal-locales/modal-locales.component.spec.ts 	renamed:    src/app/pages/previsiones/modal-locales/modal-locales.component.ts -> src/app/pages/previsiones/old- modal-locales/modal-locales.component.ts 	modified:   src/app/pages/previsiones/previsiones.page.ts
