+++ src/app/app.module.ts
@@ -6,7 +6,7 @@ import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ModalLocalesComponent } from './pages/previsiones/modal-locales/modal-locales.component';
+import { ModalLocalesComponent } from './pages/previsiones/old- modal-locales/modal-locales.component';
 
 @NgModule({
   declarations: [AppComponent],
