+++ src/app/pages/previsiones/old- modal-locales/modal-locales.component.scss
